rag_framework_base/requirements.txt
python-dotenv
setuptools
llama-index
llama-index-core
llama-index-llms-ollama
llama-index-llms-vllm
llama-index-embeddings-ollama
llama-index-embeddings-langchain
llama-index-vector-stores-milvus
llama-index-graph-stores-neo4j
llama-index-readers-file
llama-index-readers-json
llama-index-readers-mongodb
llama-index-readers-database
llama-index-readers-milvus
llama-index-readers-obsidian
llama-index-readers-whisper
llama-index-readers-graphdb-cypher
llama-index-storage-index-store-redis
llama-index-storage-docstore-mongodb
llama-index-storage-kvstore-redis
llama-index-tools-database
llama-index-tools-neo4j
llama-index-multi-modal-llms-ollama
llama-index-packs-neo4j-query-engine
llama-index-extractors-entity
llama-index-experimental
transformers
sentence-transformers
ragas
datasets
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
pyyaml
pytest
tqdm
evaluate
rouge-score
bert-score
tiktoken
langchain
nest-asyncio